Separated components.
